+++ python/lsst/ts/ess/sensors/__init__.py
@@ -1,4 +1,4 @@
-# This file is part of ts_envsensors.
+# This file is part of ts_ess_sensors.
 #
 # Developed for the Vera C. Rubin Observatory Telescope and Site Systems.
 # This product includes software developed by the LSST Project
@@ -19,4 +19,19 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from .config_jschema import CONFIG_JSCHEMA
+try:
+    from .version import *
+except ModuleNotFoundError:
+    __version__ = "?"
+
+from .command_error import *
+from .command_handler import *
+from .constants import *
+from .device_config import *
+from .response_code import *
+from .socket_server import *
+
+# Import sub modules
+from . import device
+from . import schema
+from . import sensor