+++ python/lsst/ts/ess/sensors/device/rpi_serial_hat.py
@@ -1,4 +1,4 @@
-# This file is part of ts_envsensors.
+# This file is part of ts_ess_sensors.
 #
 # Developed for the Vera C. Rubin Observatory Telescope and Site Systems.
 # This product includes software developed by the LSST Project
@@ -25,8 +25,8 @@ import asyncio
 import logging
 from typing import Callable, Tuple
 
-import RPi.GPIO as gpio  # type: ignore
-import serial  # type: ignore
+import RPi.GPIO as gpio
+import serial
 
 from .base_device import BaseDevice
 from ..response_code import ResponseCode