+++ tests/test_command_handler.py
@@ -1,4 +1,4 @@
-# This file is part of ts_envsensors.
+# This file is part of ts_ess_sensors.
 #
 # Developed for the Vera C. Rubin Observatory Telescope and Site Systems.
 # This product includes software developed by the LSST Project
@@ -23,7 +23,7 @@ import asyncio
 import logging
 import unittest
 
-from lsst.ts import envsensors
+from lsst.ts.ess import sensors
 from base_mock_test_case import BaseMockTestCase
 
 logging.basicConfig(
@@ -33,7 +33,7 @@ logging.basicConfig(
 
 class CommandHandlerTestCase(BaseMockTestCase):
     async def asyncSetUp(self):
-        self.command_handler = envsensors.CommandHandler(
+        self.command_handler = sensors.CommandHandler(
             callback=self.callback, simulation_mode=1
         )
         self.assertIsNone(self.command_handler._configuration)
@@ -42,27 +42,27 @@ class CommandHandlerTestCase(BaseMockTestCase):
         # needed because of the configuration validation. All device
         # configuration is ignored but the sensor type is not. The main
         # objective is to test multiple sensors at the same time.
-        device_config_01 = envsensors.DeviceConfig(
+        device_config_01 = sensors.DeviceConfig(
             name="Test01",
             num_channels=4,
-            dev_type=envsensors.DeviceType.FTDI,
+            dev_type=sensors.DeviceType.FTDI,
             dev_id="ABC",
-            sens_type=envsensors.SensorType.TEMPERATURE,
+            sens_type=sensors.SensorType.TEMPERATURE,
         )
-        device_config_02 = envsensors.DeviceConfig(
+        device_config_02 = sensors.DeviceConfig(
             name="Test02",
-            dev_type=envsensors.DeviceType.FTDI,
+            dev_type=sensors.DeviceType.FTDI,
             dev_id="ABC",
-            sens_type=envsensors.SensorType.HX85A,
+            sens_type=sensors.SensorType.HX85A,
         )
-        device_config_03 = envsensors.DeviceConfig(
+        device_config_03 = sensors.DeviceConfig(
             name="Test03",
-            dev_type=envsensors.DeviceType.FTDI,
+            dev_type=sensors.DeviceType.FTDI,
             dev_id="ABC",
-            sens_type=envsensors.SensorType.HX85BA,
+            sens_type=sensors.SensorType.HX85BA,
         )
         self.configuration = {
-            envsensors.Key.DEVICES: [
+            sensors.Key.DEVICES: [
                 device_config_01.as_dict(),
                 device_config_02.as_dict(),
                 device_config_03.as_dict(),
@@ -80,7 +80,7 @@ class CommandHandlerTestCase(BaseMockTestCase):
 
     def assert_response(self, response_code):
         response = self.responses.pop()
-        self.assertEqual(response[envsensors.Key.RESPONSE], response_code)
+        self.assertEqual(response[sensors.Key.RESPONSE], response_code)
 
     async def test_configure(self):
         """Test the configuration validation of the CommandHandler."""
@@ -88,105 +88,105 @@ class CommandHandlerTestCase(BaseMockTestCase):
         # Test with the configuration as set in asyncSetUp, which is a good
         # configuration.
         await self.command_handler.handle_command(
-            command=envsensors.Command.CONFIGURE, configuration=self.configuration
+            command=sensors.Command.CONFIGURE, configuration=self.configuration
         )
-        self.assert_response(envsensors.ResponseCode.OK)
+        self.assert_response(sensors.ResponseCode.OK)
         self.assertDictEqual(self.configuration, self.command_handler._configuration)
 
-        # The value for envsensors.Key.DEVICES may not be empty.
-        bad_configuration = {envsensors.Key.DEVICES: []}
+        # The value for sensors.Key.DEVICES may not be empty.
+        bad_configuration = {sensors.Key.DEVICES: []}
         await self.command_handler.handle_command(
-            command=envsensors.Command.CONFIGURE, configuration=bad_configuration
+            command=sensors.Command.CONFIGURE, configuration=bad_configuration
         )
-        self.assert_response(envsensors.ResponseCode.INVALID_CONFIGURATION)
+        self.assert_response(sensors.ResponseCode.INVALID_CONFIGURATION)
 
         # A configuration with several errors that will get tested one by one.
         bad_items = [
             {
-                # The mandatory keys envsensors.Key.CHANNELS,
-                # envsensors.Key.DEVICE_TYPE and KEY.SENSOR_TYPE are missing.
-                envsensors.Key.NAME: "Test1",
+                # The mandatory keys sensors.Key.CHANNELS,
+                # sensors.Key.DEVICE_TYPE and KEY.SENSOR_TYPE are missing.
+                sensors.Key.NAME: "Test1",
             },
             {
-                # The mandatory keys envsensors.Key.DEVICE_TYPE and
-                # envsensors.Key.SENSOR_TYPE are missing.
-                envsensors.Key.NAME: "Test1",
-                envsensors.Key.CHANNELS: 4,
+                # The mandatory keys sensors.Key.DEVICE_TYPE and
+                # sensors.Key.SENSOR_TYPE are missing.
+                sensors.Key.NAME: "Test1",
+                sensors.Key.CHANNELS: 4,
             },
             {
-                # The mandatory key envsensors.Key.SENSOR_TYPE is missing.
-                envsensors.Key.NAME: "Test1",
-                envsensors.Key.CHANNELS: 4,
-                envsensors.Key.DEVICE_TYPE: envsensors.DeviceType.FTDI,
+                # The mandatory key sensors.Key.SENSOR_TYPE is missing.
+                sensors.Key.NAME: "Test1",
+                sensors.Key.CHANNELS: 4,
+                sensors.Key.DEVICE_TYPE: sensors.DeviceType.FTDI,
             },
             {
-                # For envsensors.DeviceType.FTDI the key
-                # envsensors.Key.FTDI_ID is mandatory.
-                envsensors.Key.NAME: "Test1",
-                envsensors.Key.CHANNELS: 4,
-                envsensors.Key.DEVICE_TYPE: envsensors.DeviceType.FTDI,
+                # For sensors.DeviceType.FTDI the key
+                # sensors.Key.FTDI_ID is mandatory.
+                sensors.Key.NAME: "Test1",
+                sensors.Key.CHANNELS: 4,
+                sensors.Key.DEVICE_TYPE: sensors.DeviceType.FTDI,
                 "id": "ABC",
-                envsensors.Key.SENSOR_TYPE: envsensors.SensorType.TEMPERATURE,
+                sensors.Key.SENSOR_TYPE: sensors.SensorType.TEMPERATURE,
             },
             {
-                # For envsensors.DeviceType.SERIAL the key
-                # envsensors.Key.SERIAL_PORT is mandatory.
-                envsensors.Key.NAME: "Test1",
-                envsensors.Key.CHANNELS: 4,
-                envsensors.Key.DEVICE_TYPE: envsensors.DeviceType.SERIAL,
+                # For sensors.DeviceType.SERIAL the key
+                # sensors.Key.SERIAL_PORT is mandatory.
+                sensors.Key.NAME: "Test1",
+                sensors.Key.CHANNELS: 4,
+                sensors.Key.DEVICE_TYPE: sensors.DeviceType.SERIAL,
                 "port": "ABC",
-                envsensors.Key.SENSOR_TYPE: envsensors.SensorType.TEMPERATURE,
+                sensors.Key.SENSOR_TYPE: sensors.SensorType.TEMPERATURE,
             },
             {
-                # Wrong value for envsensors.Key.SENSOR_TYPE.
-                envsensors.Key.NAME: "Test1",
-                envsensors.Key.CHANNELS: 4,
-                envsensors.Key.DEVICE_TYPE: envsensors.DeviceType.SERIAL,
-                envsensors.Key.SERIAL_PORT: "ABC",
-                envsensors.Key.SENSOR_TYPE: "Temp",
+                # Wrong value for sensors.Key.SENSOR_TYPE.
+                sensors.Key.NAME: "Test1",
+                sensors.Key.CHANNELS: 4,
+                sensors.Key.DEVICE_TYPE: sensors.DeviceType.SERIAL,
+                sensors.Key.SERIAL_PORT: "ABC",
+                sensors.Key.SENSOR_TYPE: "Temp",
             },
         ]
         for bad_item in bad_items:
-            bad_configuration = {envsensors.Key.DEVICES: [bad_item]}
+            bad_configuration = {sensors.Key.DEVICES: [bad_item]}
             await self.command_handler.handle_command(
-                command=envsensors.Command.CONFIGURE, configuration=bad_configuration
+                command=sensors.Command.CONFIGURE, configuration=bad_configuration
             )
-            self.assert_response(envsensors.ResponseCode.INVALID_CONFIGURATION)
+            self.assert_response(sensors.ResponseCode.INVALID_CONFIGURATION)
 
     async def test_start(self):
         """Test handling of the start command."""
-        await self.command_handler.handle_command(command=envsensors.Command.START)
-        self.assert_response(envsensors.ResponseCode.NOT_CONFIGURED)
+        await self.command_handler.handle_command(command=sensors.Command.START)
+        self.assert_response(sensors.ResponseCode.NOT_CONFIGURED)
         self.assertIsNone(self.command_handler._configuration)
         self.assertFalse(self.command_handler._started)
 
         await self.command_handler.handle_command(
-            command=envsensors.Command.CONFIGURE, configuration=self.configuration
+            command=sensors.Command.CONFIGURE, configuration=self.configuration
         )
-        self.assert_response(envsensors.ResponseCode.OK)
+        self.assert_response(sensors.ResponseCode.OK)
         self.assertDictEqual(self.configuration, self.command_handler._configuration)
-        await self.command_handler.handle_command(command=envsensors.Command.START)
-        self.assert_response(envsensors.ResponseCode.OK)
+        await self.command_handler.handle_command(command=sensors.Command.START)
+        self.assert_response(sensors.ResponseCode.OK)
         self.assertTrue(self.command_handler._started)
 
     async def test_stop(self):
         """Test handling of the stop command."""
-        await self.command_handler.handle_command(command=envsensors.Command.STOP)
-        self.assert_response(envsensors.ResponseCode.NOT_STARTED)
+        await self.command_handler.handle_command(command=sensors.Command.STOP)
+        self.assert_response(sensors.ResponseCode.NOT_STARTED)
         self.assertIsNone(self.command_handler._configuration)
         self.assertFalse(self.command_handler._started)
 
         await self.command_handler.handle_command(
-            command=envsensors.Command.CONFIGURE, configuration=self.configuration
+            command=sensors.Command.CONFIGURE, configuration=self.configuration
         )
-        self.assert_response(envsensors.ResponseCode.OK)
+        self.assert_response(sensors.ResponseCode.OK)
         self.assertDictEqual(self.configuration, self.command_handler._configuration)
-        await self.command_handler.handle_command(command=envsensors.Command.START)
-        self.assert_response(envsensors.ResponseCode.OK)
+        await self.command_handler.handle_command(command=sensors.Command.START)
+        self.assert_response(sensors.ResponseCode.OK)
         self.assertTrue(self.command_handler._started)
 
-        await self.command_handler.handle_command(command=envsensors.Command.STOP)
-        self.assert_response(envsensors.ResponseCode.OK)
+        await self.command_handler.handle_command(command=sensors.Command.STOP)
+        self.assert_response(sensors.ResponseCode.OK)
         # Give time to the telemetry_task to get cancelled.
         await asyncio.sleep(0.5)
         self.assertFalse(self.command_handler._started)
@@ -194,12 +194,12 @@ class CommandHandlerTestCase(BaseMockTestCase):
     async def test_get_telemetry(self):
         """Test handling of telemetry."""
         await self.command_handler.handle_command(
-            command=envsensors.Command.CONFIGURE, configuration=self.configuration
+            command=sensors.Command.CONFIGURE, configuration=self.configuration
         )
-        self.assert_response(envsensors.ResponseCode.OK)
+        self.assert_response(sensors.ResponseCode.OK)
         self.assertDictEqual(self.configuration, self.command_handler._configuration)
-        await self.command_handler.handle_command(command=envsensors.Command.START)
-        self.assert_response(envsensors.ResponseCode.OK)
+        await self.command_handler.handle_command(command=sensors.Command.START)
+        self.assert_response(sensors.ResponseCode.OK)
         self.assertTrue(self.command_handler._started)
 
         # Give some time to the mock sensor to produce data
@@ -209,23 +209,23 @@ class CommandHandlerTestCase(BaseMockTestCase):
         devices_names_checked = set()
         while len(devices_names_checked) != len(self.device_configs):
             reply = self.responses.pop()
-            self.name = reply[envsensors.Key.TELEMETRY][0]
+            self.name = reply[sensors.Key.TELEMETRY][0]
             devices_names_checked.add(self.name)
             device_config = self.device_configs[self.name]
             self.disconnected_channel = None
             self.missed_channels = 0
             self.in_error_state = False
-            reply_to_check = reply[envsensors.Key.TELEMETRY]
-            if device_config.sens_type == envsensors.SensorType.TEMPERATURE:
+            reply_to_check = reply[sensors.Key.TELEMETRY]
+            if device_config.sens_type == sensors.SensorType.TEMPERATURE:
                 self.num_channels = device_config.num_channels
                 self.check_temperature_reply(reply_to_check)
-            elif device_config.sens_type == envsensors.SensorType.HX85A:
+            elif device_config.sens_type == sensors.SensorType.HX85A:
                 self.check_hx85a_reply(reply_to_check)
             else:
                 self.check_hx85ba_reply(reply_to_check)
 
-        await self.command_handler.handle_command(command=envsensors.Command.STOP)
-        self.assert_response(envsensors.ResponseCode.OK)
+        await self.command_handler.handle_command(command=sensors.Command.STOP)
+        self.assert_response(sensors.ResponseCode.OK)
         # Give time to the telemetry_task to get cancelled.
         await asyncio.sleep(0.5)
         self.assertFalse(self.command_handler._started)